refactor: cube response deserialization and graphql objects for projects. stub out organizations and collections for the time being.
